Label sub unit radio buttons
Providing a label for the second set of radio button input's. With just one label sighted users have to infer what the second set of buttons designation is from the radio button choices. Non-visual access is hindered as screen readers will not read off what the second set of radio buttons is for. Recommending a new label as opposed to a hidden aria-label